fix(extensions): second review wave — digest scope, atomic write, error framing
Follow-up to the cubic review of the previous fixes (several were fallout
from those changes):

- Digest no longer skips dot-prefixed files (a go:embed 'all:' pattern can
  ship hidden assets, whose change must invalidate the composition); the
  earlier .DS_Store concern was moot since composition.json is gitignored
  and per-run. Instead the published-surface sweep excludes only *_test.go
  — tests are in neither the compile-against surface nor the binary, so a
  test edit must not force a rebuild.
- Manifest write is now temp-file + atomic rename: rename replaces the
  path without following a symlink and has no check-then-write TOCTOU, so
  the earlier fail-open Lstat guard is gone.
- drift also fails on an UNTRACKED committed-tier manifest (git diff
  ignores untracked), catching a first-party extension whose manifest was
  never git-added.
- Semantic validation errors (a bad version, an out-of-vocabulary tool
  scope) now carry position only, not the 'declare literal values'
  prescription, which applies to shape errors; the Tier constants and docs
  describe the declared tier/scope as a REQUEST, not effective behavior.
- Docs: the §1 capability summary lists Tools; the literal requirement
  notes the published extension constants are accepted; a missing 'that'.

Split unitmanifest.go (the AST reader moved to astreader.go) to clear the
file-length gate. Tests: _test.go excluded / dot-asset hashed, and the
existing repeated-const, multiple-New, whitespace-version cases.
